Correct build for LLVM 16
Do bare minimum to compile with LLVM 16. Further cleanup
of warnings recommended.

+++ include/os/windows/spl/sys/kmem_impl.h
@@ -393,7 +393,7 @@ struct kmem_cache {
 	int		(*cache_constructor)(void *, void *, int);
 	void		(*cache_destructor)(void *, void *);
 	void		(*cache_reclaim)(void *);
-	kmem_cbrc_t	(*cache_move)(void *, void *, uint32_t, void *);
+	kmem_cbrc_t	(*cache_move)(void *, void *, size_t, void *);
 	void		*cache_private;	/* opaque arg to callbacks */
 	vmem_t		*cache_arena;	/* vmem source for slabs */
 	int		cache_cflags;	/* cache creation flags */
